V2

+++ src/Controller/TaskController.php
@@ -6,19 +6,21 @@ use App\Entity\Task;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\Exception\ORMException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class TaskController extends AbstractController
 {
 
 
-    public function __construct(private readonly EntityManagerInterface $em, private readonly TaskRepository $taskRepository)
-    {
+    public function __construct(
+        private readonly EntityManagerInterface $em,
+        private readonly TaskRepository $taskRepository
+    ) {
     }
 
     #[Route('/tasks', name: 'task_list')]
@@ -28,18 +30,17 @@ class TaskController extends AbstractController
     }
 
     /**
-     * @throws ORMException
      */
     #[Route('/tasks/create', name: 'task_create')]
     public function createAction(Request $request): RedirectResponse|Response
     {
         $task = new Task();
-        $form = $this->createForm(TaskType::class, $task);
+        $task->setUser($this->getUser());
 
+        $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-
             $this->em->persist($task);
             $this->em->flush();
 
@@ -54,6 +55,12 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/edit', name: 'task_edit')]
     public function editAction(Task $task, Request $request): RedirectResponse|Response
     {
+        $user = $this->getUser();
+
+        if ($task->getUser() !== $user) {
+            throw new AccessDeniedException('Vous ne pouvez pas modifier une tâche qui ne vous appartient pas.');
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -75,6 +82,14 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTaskAction(Task $task): RedirectResponse
     {
+        $user = $this->getUser();
+
+        if ($task->getUser() !== $user) {
+            throw new AccessDeniedException(
+                'Vous ne pouvez pas indiquer une tâche comme terminée qui ne vous appartient pas.'
+            );
+        }
+
         $task->toggle(!$task->isDone());
         $this->em->flush();
 
@@ -86,6 +101,13 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTaskAction(Task $task): RedirectResponse
     {
+        $user = $this->getUser();
+        $taskUser = $task->getUser();
+
+        if (($taskUser !== $user) && ($taskUser->getUsername() !== 'Anonyme' || !$this->isGranted('ROLE_ADMIN'))) {
+            throw new AccessDeniedException('Vous ne pouvez pas supprimer cette tâche.');
+        }
+
         $this->em->remove($task);
         $this->em->flush();
